+++ resources/folder.rb
@@ -1,6 +1,7 @@
 # To learn more about Custom Resources, see https://docs.chef.io/custom_resources.html
 property :host,           String,   default: 'localhost'
 property :port,           Integer,  default: 3000
+property :url_path_prefix, String
 property :admin_user,     String,   default: 'admin'
 property :admin_password, String,   default: 'admin'
 property :folder,         Hash,     default: {}
@@ -17,6 +18,7 @@ action :create do
   grafana_options = {
     host: new_resource.host,
     port: new_resource.port,
+    url_path_prefix: new_resource.url_path_prefix,
     user: new_resource.admin_user,
     password: new_resource.admin_password,
     auth_proxy_header: new_resource.auth_proxy_header,
@@ -51,6 +53,7 @@ action :update do
   grafana_options = {
     host: new_resource.host,
     port: new_resource.port,
+    url_path_prefix: new_resource.url_path_prefix,
     user: new_resource.admin_user,
     password: new_resource.admin_password,
     auth_proxy_header: new_resource.auth_proxy_header,
@@ -92,6 +95,7 @@ action :delete do
   grafana_options = {
     host: new_resource.host,
     port: new_resource.port,
+    url_path_prefix: new_resource.url_path_prefix,
     user: new_resource.admin_user,
     password: new_resource.admin_password,
     auth_proxy_header: new_resource.auth_proxy_header,