+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This file is used to list changes made in each version of grafana.
 
+## Unreleased
+
+- Added support for HTTPS endpoint
+- Added support for endpoints with url_path_prefix
+- Fixed url_path_prefix property definition in resources
+
 ## 9.3.0 - 2020-08-19
 
 - Added support for use of config_writer to only generate config file when grafana is not present on host